0.7.0
updated client/src/pages/savedbooks.js to use graphql by importing the useQuery, useMutations hooks, and the queries and mutations from util. which delete the book with del book function by the bookId updated the if conditional on line 43 to use loading as its an option now from the useQuery method.